test(cron): from Jest to test
